eval(tkde): fix result-heavy generator overflow and freeze contracts v1.1
The frozen Result-heavy Dataset Generator evaluated row_id * 104729 as a
PostgreSQL int4 and overflowed deterministically from row_id 20504, so
ON_ERROR_STOP aborted the whole generator and no deployment could ever
materialize reporting.final_v5_result_heavy. Promote row_id to bigint before
the multiplication, which is the 64-bit arithmetic the independent Oracle
always specified, and re-freeze the corrected bytes as contract release
final-v5-contracts-v1.1. contracts/AMENDMENT-v1.1.md records the defect, the
absent publication impact, and every regenerated transitive digest.

Add a Contract-to-Runtime Bridge so the Contract Index and the machine
contracts are the runtime source of truth: it loads and re-derives every index
digest from embedded contract bytes, revalidates the hash-locked protocol
references, looks cells up, renders the frozen query pair, binds Dataset,
Catalog and Publication against the live deployment, loads and independently
recomputes each Oracle Manifest, normalizes Direct and BDG results through one
shared normalizer, compares expected against actual, and reports profile
completeness fail-closed. The Artifact Adapter and its capability matrix are
now derived from that Bridge instead of a second hand-maintained table.

Wire the final_v5_result_heavy Product and the final-v5-result-heavy-v1
Publication into the live stack: the deployment executes the contract-indexed
generator itself, compiles the Publication with its own offline compiler, and
installs generated sidecar/dictionary/manifest digests plus a regenerated
source attestation into config/catalog.yaml. Two independent fresh deployments
produce byte-identical datasets, and all six Artifact shapes agree with the
independent Oracle on row count, normalized schema digest and canonical
logical-result digest.

artifact stays false and the capability report stays 6/9. docs C11 records the
one remaining blocker: the live HOT artifacts now total 233.4 MiB against the
Gateway's 160 MiB activation boundary, which is a production limit this change
deliberately does not bypass. No Campaign ID, no submission commit, no formal
Campaign, and no paper number was produced.

+++ evaluation/final-v5-wsl2/sql/datasets/benchmark-v1-generate.sql
@@ -78,7 +78,7 @@ SELECT row_id,
            + (((row_id - 1) % 1000) * interval '1 microsecond'),
        'artifact-row-' || row_id::text,
        1 + ((row_id - 1) % 10000),
-       ((row_id * 104729) % 10000000)::numeric / 10000,
+       ((row_id::bigint * 104729) % 10000000)::numeric / 10000,
        CASE WHEN (row_id % 11) = 0 THEN -1 ELSE 1 END
            * (((row_id * 37) % 1000000)::numeric / 100),
        date '2020-01-01' + ((row_id - 1 + 31) % 3653)::integer,
